base/serializers.py
from hashid_field.rest import HashidSerializerCharField
from rest_framework import serializers
from django.contrib.auth.models import User
from rest_framework_simplejwt.tokens import RefreshToken
from .models import Product, Order, OrderItem, ShippingAddress, Review, Feature, FeatureType, FeatureStatus


class UserSerializer(serializers.ModelSerializer):
    name = serializers.SerializerMethodField(read_only=True)
    _id = serializers.SerializerMethodField(read_only=True)
    isAdmin = serializers.SerializerMethodField(read_only=True)

    class Meta:
        model = User
        fields = ['id', '_id', 'username', 'email', 'name', 'isAdmin']

    def get__id(self, obj):
        return obj.id

    def get_isAdmin(self, obj):
        return obj.is_staff

    def get_name(self, obj):
        name = obj.first_name
        if name == '':
            name = obj.email

        return name


class UserSerializerWithToken(UserSerializer):
    token = serializers.SerializerMethodField(read_only=True)

    class Meta:
        model = User
        fields = ['id', '_id', 'username', 'email', 'name', 'isAdmin', 'token']

    def get_token(self, obj):
        token = RefreshToken.for_user(obj)
        return str(token.access_token)


class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = '__all__'


class FeatureTypeSerializer(serializers.ModelSerializer):
    class Meta:
        model = FeatureType
        fields = ('id', 'name', 'color')


class FeatureStatusSerializer(serializers.ModelSerializer):
    class Meta:
        model = FeatureStatus
        fields = ('id', 'name', 'color')


class FeatureSerializer(serializers.ModelSerializer):
    id = serializers.CharField(read_only=True)

    class Meta:
        model = Feature
        fields = ('id', 'name', 'slug', 'description', 'votes', 'type', 'status', 'createdAt')


class ProductSerializer(serializers.ModelSerializer):
    reviews = serializers.SerializerMethodField(read_only=True)
    id = serializers.CharField(read_only=True)
    features = serializers.SerializerMethodField(read_only=True)
    featureTypes = serializers.SerializerMethodField(read_only=True)
    featureStatuses = serializers.SerializerMethodField(read_only=True)

    class Meta:
        model = Product
        fields = ('id', 'name', 'features', 'description', 'logo', 'primaryColor', 'secondaryColor', 'websiteUrl',
                  'reviews', 'featureTypes', 'featureStatuses')

    def get_reviews(self, obj):
        reviews = obj.review_set.all()
        serializer = ReviewSerializer(reviews, many=True)
        return serializer.data

    def get_features(self, obj):
        features = Feature.objects.filter(product=obj.id, public=True)
        serializer = FeatureSerializer(features, many=True)
        return serializer.data

    def get_featureTypes(self, obj):
        feature_types = FeatureType.objects.all()
        serializer = FeatureTypeSerializer(feature_types, many=True)
        return serializer.data

    def get_featureStatuses(self, obj):
        feature_statuses = FeatureStatus.objects.all()
        serializer = FeatureTypeSerializer(feature_statuses, many=True)
        return serializer.data


class ShippingAddressSerializer(serializers.ModelSerializer):
    class Meta:
        model = ShippingAddress
        fields = '__all__'


class OrderItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = OrderItem
        fields = '__all__'


class OrderSerializer(serializers.ModelSerializer):
    orderItems = serializers.SerializerMethodField(read_only=True)
    shippingAddress = serializers.SerializerMethodField(read_only=True)
    user = serializers.SerializerMethodField(read_only=True)

    class Meta:
        model = Order
        fields = '__all__'

    def get_orderItems(self, obj):
        items = obj.orderitem_set.all()
        serializer = OrderItemSerializer(items, many=True)
        return serializer.data

    def get_shippingAddress(self, obj):
        try:
            address = ShippingAddressSerializer(
                obj.shippingaddress, many=False).data
        except:
            address = False
        return address

    def get_user(self, obj):
        user = obj.user
        serializer = UserSerializer(user, many=False)
        return serializer.data
